ExprParser

Description

Float Parsers

parseExprF

Arguments

:: String
input string for attempted parse

-> Expr Float
resulting expression in Expr type of Float

Attempts to parses a string into an expression of type Float. If invalid expression throws error.

exprF :: Parser (Expr Float)

Parses each operator individually

Unary Operations

unaryOpsF :: Parser (Expr Float)

Parses all unary operators

binaryOpsF :: Parser (Expr Float)

Parses all binary operators

parseVarF :: Parser (Expr Float)

Parses Variables

parseConstF :: Parser (Expr Float)

Parses Constants of Float

parseSinF :: Parser (Expr Float)

Parses Sin function of Float

parseCosF :: Parser (Expr Float)

Parses Cos function of Float

parseLnF :: Parser (Expr Float)

Parses natural logarithm function of Float

parseEF :: Parser (Expr Float)

Parses natural exponential function of Float

Binary Operations

parseMultF :: Parser (Expr Float)

Parses binary multiplication

parseAddF :: Parser (Expr Float)

Parses binary addition

Double Parsers

parseExprD

Arguments

:: String
input string for attempted parse

-> Expr Double
resulting expression in Expr type of Double

Attempts to parses a string into an expression of type Double. ID invalid expression throws error.

exprD :: Parser (Expr Double)

Parses each operator individually

Unary Operations

unaryOpsD :: Parser (Expr Double)

Parses all unary operators

binaryOpsD :: Parser (Expr Double)

Parses all binary operators

parseVarD :: Parser (Expr Double)

Parses Variables

parseConstD :: Parser (Expr Double)

Parses Constants of Double

parseSinD :: Parser (Expr Double)

Parses Sin Function of Double

parseCosD :: Parser (Expr Double)

Parses Cos Function of Double

parseLnD :: Parser (Expr Double)

Parses natural logarithm Function of Double

parseED :: Parser (Expr Double)

Parses natural exponential Function of Double

Binary Operations

parseMultD :: Parser (Expr Double)

Parses binary multiplication

parseAddD :: Parser (Expr Double)

Parses binary addition

Miscellaneous Parsers

parens :: Parser a -> Parser a

symbol :: String -> Parser String

digits :: Parser String

negDigits :: Parser String

integer :: Parser Integer

decimalDigits :: Parser String

negDecimalDigits :: Parser String

double :: Parser Double

float :: Parser Float